packages/store/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prlens-store"
version = "0.1.8"
description = "Pluggable review history store for prlens"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "PyGithub>=2.1",
]

[project.optional-dependencies]
sqlite = []
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "black>=24.0",
    "flake8>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
